# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ERB System API Documentation",
        "description": "Tài liệu đặc tả và kiểm thử RESTful API - Hệ thống quản trị ERB",
        "contact": {
            "email": "admin@erb.vn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/skus": {
            "get": {
                "tags": [
                    "SKUs"
                ],
                "summary": "Lấy danh sách sản phẩm (SKU)",
                "description": "Danh sách SKU hỗ trợ tìm kiếm theo mã/tên, lọc theo nhóm hàng, trạng thái, có serial và phân trang.",
                "operationId": "73c6885b75631392e1877fae77796832",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Số bản ghi/trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Tìm theo mã hoặc tên SKU",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Lọc theo nhóm sản phẩm",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "has_serial",
                        "in": "query",
                        "description": "Quản lý serial (1: Có, 0: Không)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Trạng thái (1: Hoạt động, 0: Ngừng)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "code": {
                                                                "type": "string",
                                                                "example": "SP001"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Áo thun nam"
                                                            },
                                                            "unit": {
                                                                "type": "string",
                                                                "example": "cái"
                                                            },
                                                            "category": {
                                                                "type": "string",
                                                                "example": "Thời trang",
                                                                "nullable": true
                                                            },
                                                            "has_serial": {
                                                                "type": "boolean",
                                                                "example": false
                                                            },
                                                            "status": {
                                                                "type": "integer",
                                                                "example": 1
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực tài khoản"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SKUs"
                ],
                "summary": "Tạo SKU mới",
                "description": "Tạo sản phẩm. Mã SKU (code) là duy nhất toàn hệ thống.",
                "operationId": "bffb584c125377fcaf24548abdaf69c1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "name",
                                    "unit"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "SP001"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Áo thun nam"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "cái"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Thời trang",
                                        "nullable": true
                                    },
                                    "has_serial": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "status": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ],
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo SKU thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "401": {
                        "description": "Chưa xác thực tài khoản"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/skus/{id}": {
            "get": {
                "tags": [
                    "SKUs"
                ],
                "summary": "Xem chi tiết SKU",
                "operationId": "15014f0b9bc327c85638f1f492921405",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy SKU"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "SKUs"
                ],
                "summary": "Cập nhật SKU",
                "operationId": "2672f3e8654cfbe8fda3906ea1378a36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "SP001"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Áo thun nam (mới)"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "cái"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Thời trang",
                                        "nullable": true
                                    },
                                    "has_serial": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "status": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ],
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "404": {
                        "description": "Không tìm thấy SKU"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "SKUs"
                ],
                "summary": "Xóa SKU",
                "description": "Chỉ xóa được khi SKU chưa phát sinh tồn/giao dịch. Khuyến nghị dùng Ngừng hoạt động thay vì xóa.",
                "operationId": "4384a26f2e2dfebf8f432bb59f2062f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy SKU"
                    },
                    "409": {
                        "description": "SKU đang được sử dụng, không thể xóa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/skus/{id}/disable": {
            "patch": {
                "tags": [
                    "SKUs"
                ],
                "summary": "Ngừng hoạt động SKU",
                "operationId": "4685ead7aecc793d6e08b20d7916a242",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đã ngừng hoạt động"
                    },
                    "404": {
                        "description": "Không tìm thấy SKU"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/skus/{id}/enable": {
            "patch": {
                "tags": [
                    "SKUs"
                ],
                "summary": "Kích hoạt lại SKU",
                "operationId": "4ec3c65abd991763f9d189994f472762",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đã kích hoạt"
                    },
                    "404": {
                        "description": "Không tìm thấy SKU"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập JWT Access Token (dạng: Bearer <token>) để thực hiện các yêu cầu cần xác thực.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}